Add testing
This commit adds testing (through the `jest` package) and, as necessary
step, changes the file structure of the project to work with jest's
syntax of `require([path_to_file])`.

Testing support has been added for the primary function,
`create_mine_guide()`, as well as its two helper functions,
`sort_nodeS_descending()` and `get_legal_directions()`.
