+++ package.json
@@ -22,7 +22,8 @@
   },
   "scripts": {
     "start": "node -r esm --no-warnings ./index.js",
-    "mine": "node -r esm --no-warnings ./index.js"
+    "mine": "node -r esm --no-warnings ./index.js",
+    "test": "jest"
   },
   "main": "./index.js",
   "engines": {
@@ -32,5 +33,8 @@
   "dependencies": {
     "esm": "^3.2.25",
     "require-directory": "^2.1.1"
+  },
+  "devDependencies": {
+    "jest": "^27.0.6"
   }
 }
