+++ src/tests/create_mine_guide.test.js
@@ -1,0 +1,46 @@
+const create_mine_guide = require("../create_mine_guide");
+
+const titan = [
+  [5, 2, 0, 9],
+  [6, 6, 1, 3],
+  [2, 6, 0, 2],
+  [8, 7, 0, 9]
+];
+const titan_best_starting_row = 0;
+const titan_best_path = [1, 0, -1];
+
+const luna = [
+  [1, 2, 2, 4, 8, 3, 5, 9, 3, 7],
+  [6, 9, 8, 7, 6, 3, 1, 5, 7, 8],
+  [6, 1, 0, 1, 1, 9, 2, 8, 2, 1],
+  [1, 5, 4, 7, 6, 5, 0, 4, 0, 9],
+  [6, 5, 3, 5, 7, 1, 7, 7, 6, 5],
+  [3, 2, 7, 8, 5, 1, 9, 1, 9, 6],
+  [6, 8, 4, 1, 7, 6, 2, 1, 6, 1],
+  [5, 6, 1, 9, 7, 2, 2, 1, 0, 6],
+  [0, 4, 9, 5, 4, 8, 8, 7, 2, 4],
+  [9, 9, 0, 1, 8, 9, 5, 2, 0, 9]
+];
+const luna_best_starting_row = 9;
+const luna_best_path = [0, -1, 0, 1, -1, 1, -1, 0, 1];
+
+
+describe("Extract optimal mine info", () => {
+  const titan_best_path_info = create_mine_guide(titan);
+  const luna_best_path_info = create_mine_guide(luna);
+
+  test("determine best starting row for titan", () => {
+    expect(titan_best_path_info.best_start_row).toBe(titan_best_starting_row);
+  });
+  test("determine best starting row path for titan", () => {
+    expect(titan_best_path_info.best_path).toStrictEqual(titan_best_path);
+  });
+
+  test("determine best starting row for luna", () => {
+    expect(luna_best_path_info.best_start_row).toBe(luna_best_starting_row);
+  });
+  test("determine best starting row path for luna", () => {
+    expect(luna_best_path_info.best_path).toStrictEqual(luna_best_path);
+  });
+
+});
